Conflicts:
	proxy/common/src/main/java/eu/clarussecure/proxy/spi/Capabilities.java
	proxy/common/src/main/java/eu/clarussecure/proxy/spi/protocol/Protocol.java
	proxy/common/src/main/java/eu/clarussecure/proxy/spi/protocol/ProtocolService.java
	proxy/common/src/main/java/eu/clarussecure/proxy/spi/security/policy/MyErrorHandler.java
	proxy/main/src/main/java/eu/clarussecure/proxy/protection/ProtectionModuleLoader.java
	proxy/main/src/main/java/eu/clarussecure/proxy/protocol/ProtocolLoader.java
	proxy/plugins/pgsql/src/main/java/eu/clarussecure/proxy/protocol/plugins/pgsql/message/AuthenticationHandler.java
	proxy/plugins/pgsql/src/main/java/eu/clarussecure/proxy/protocol/plugins/pgsql/message/sql/NoopEventProcessor.java
	proxy/plugins/pgsql/src/main/java/eu/clarussecure/proxy/protocol/plugins/pgsql/message/sql/PgsqlEventProcessor.java
	proxy/plugins/pgsql/src/main/java/eu/clarussecure/proxy/protocol/plugins/pgsql/raw/handler/PgsqlRawMessage.java
	proxy/plugins/pgsql/src/test/java/eu/clarussecure/proxy/protocol/plugins/pgsql/SessionHealth.java
	proxy/plugins/pgsql/src/test/java/eu/clarussecure/proxy/protocol/plugins/pgsql/SessionIT.java
	proxy/plugins/pgsql/src/test/java/eu/clarussecure/proxy/protocol/plugins/pgsql/TestUtils.java

+++ proxy/plugins/pgsql/src/main/java/eu/clarussecure/proxy/protocol/plugins/pgsql/message/parser/PgsqlRowDescriptionMessageParser.java
@@ -21,7 +21,7 @@ public class PgsqlRowDescriptionMessageParser implements PgsqlMessageParser<Pgsq
             CString name = PgsqlUtilities.getCString(content);
             int tableOID = content.readInt();
             short columnNumber = content.readShort();
-            long typeOID = (long) content.readInt() & 0xffffffffl;
+            long typeOID = content.readInt() & 0xffffffffl;
             short typeSize = content.readShort();
             int typeModifier = content.readInt();
             short format = content.readShort();
